Move copy event into utils
